backend/requirements.txt
# Dependencies needed only for GCS upload and basic function execution
functions-framework>=3.0
google-cloud-storage
python-dotenv
Pillow # Keep Pillow, might be useful for image handling later

